Repaired Backend issues

+++ Backend/Filmer-Backend/src/main/java/com/filmer/filmerbackend/Entities/UserSensitiveData.java
@@ -15,7 +15,7 @@ public class UserSensitiveData {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id_user_sensitive_data;
+    private int id_user_sensitive_data;
 
     @Column(nullable = false, length = 45)
     private String email;
